Added all files from the stock website template
All files from the stock website template were downloaded straight from the original providers website at https://templatesjungle.gumroad.com/l/wingtea-tea-shop-bootstrap-website-template . No changes were made and nothing was edited. Refer to this when needing to look at the standard files and/or debugging. ALL OF THE FILES ARE UNORGANIZED AND NOT IN THEIR FOLDERS